Reformat vector index related logs

+++ src/VectorIndex/HNSWPQ.cpp
@@ -134,7 +134,7 @@ void * HNSWpq::convertInnerBitMap(GeneralBitMapPtr outerBitMap)
     return new_map;
 }
 
-BinaryPtr HNSWpq::convertStructToBinary(uint8_t * index_data, size_t written_size)
+BinaryPtr HNSWpq::convertStructToBinary(uint8_t * index_data, uint64_t written_size)
 {
     BinaryPtr serial_index = std::make_shared<Binary>();
     serial_index->data = index_data;
